+++ R/clusters.R
@@ -774,6 +774,8 @@ db_cluster_events <- function(cluster_id,
 #' @param polling_interval Number of seconds to wait between status checks
 #' @inheritParams auth_params
 #' @inheritParams db_cluster_edit
+#' @param silent Boolean (default: `FALSE`), will emit cluster state progress
+#' if `TRUE`.
 #'
 #' @details Get information regarding a Databricks cluster. If the cluster is
 #' inactive it will be started and wait until the cluster is active.
@@ -786,20 +788,47 @@ db_cluster_events <- function(cluster_id,
 #' @return `db_cluster_get()`
 #' @export
 get_and_start_cluster <- function(cluster_id, polling_interval = 5,
-                                  host = db_host(), token = db_token()) {
+                                  host = db_host(), token = db_token(),
+                                  silent = FALSE) {
+
+
+
 
   # get cluster status
   cluster_status <- db_cluster_get(cluster_id = cluster_id, host = host, token = token)
 
+  if (!silent) {
+    msg <- "{.header Checking cluster:} {.emph '{cluster_id}'}"
+    msg_done <- "{.header Checking cluster:} {.emph {cluster_status$cluster_name}}"
+    cli::cli_progress_step(msg, msg_done)
+  }
+
   # if the cluster isn't running, start it
-  if (!cluster_status$state %in% c("RUNNING", "PENDING")) {
+  if (!cluster_status$state %in% c("RUNNING", "PENDING", "RESIZING", "STARTING", "RESTARTING")) {
     db_cluster_start(cluster_id = cluster_id, host = host, token = token)
   }
 
   # wait for cluster to become active
+  if (!silent) {
+    msg <- "{.header Attaching to cluster:} {.emph [{cluster_status$state}] '{cluster_status$state_message}'}"
+    msg_done <- "{.header Attached to cluster}"
+    msg_failed <- "{.header Cluster entered [{cluster_status$state}] state}"
+    cli::cli_progress_step(msg, msg_done, msg_failed)
+  }
+
   while (cluster_status$state != "RUNNING") {
     Sys.sleep(polling_interval)
     cluster_status <- db_cluster_get(cluster_id = cluster_id, host = host, token = token)
+    if (!silent) cli::cli_progress_update()
+    if (cluster_status$state %in% c("TERMINATED", "TERMINATING")) {
+      if (!silent) cli::cli_progress_done(result = "failed")
+      break
+    }
+  }
+
+  if (!silent) {
+    cli::cli_progress_done()
+    cli::cli_end()
   }
 
   cluster_status