+++ R/execution-context.R
@@ -165,9 +165,67 @@ db_context_command_run <- function(cluster_id,
   )
 
     if (perform_request) {
-    db_perform_request(req)
+    res <- db_perform_request(req)
+    list(id = res$id, language = language)
   } else {
     req
+  }
+
+}
+
+#' Run a Command and Wait For Results
+#'
+#' @param parse_result Boolean, determines if results are parsed automatically.
+#' @inheritParams db_context_command_run
+#'
+#' @family Execution Context API
+#'
+#' @export
+db_context_command_run_and_wait <- function(cluster_id,
+                                            context_id,
+                                            language = c("python", "sql", "scala", "r"),
+                                            command = NULL,
+                                            command_file = NULL,
+                                            options = list(),
+                                            parse_result = TRUE,
+                                            host = db_host(), token = db_token()) {
+
+  stopifnot(is.logical(parse_result))
+
+  command <- db_context_command_run(
+    cluster_id = cluster_id,
+    context_id = context_id,
+    language = language,
+    command = command,
+    command_file = command_file,
+    options = options,
+    host = host,
+    token = token
+  )
+
+  command_status <- db_context_command_status(
+    cluster_id = cluster_id,
+    context_id = context_id,
+    command_id = command$id,
+    host = host,
+    token = token
+  )
+
+  while (command_status$status %in% c("Running", "Queued")) {
+    Sys.sleep(0.5)
+    command_status <- db_context_command_status(
+      cluster_id = cluster_id,
+      context_id = context_id,
+      command_id = command$id,
+      host = host,
+      token = token
+    )
+  }
+
+  if (parse_result) {
+    db_context_command_parse(command_status, language = language)
+  } else {
+    command_status
   }
 
 }
@@ -247,4 +305,84 @@ db_context_command_cancel <- function(cluster_id,
     req
   }
 
+}
+
+#' Parse Command Results
+#'
+#' @param x command output from `db_context_command_status` or
+#' `db_context_manager`'s `cmd_run`
+#' @param language
+#'
+#' @family Execution Context API
+#'
+#' @return command results
+#' @keywords internal
+db_context_command_parse <- function(x, language = c("r", "py", "scala", "sql")) {
+
+  language <- match.arg(language)
+
+  if (x$results$resultType == "error") {
+    cli_alert_danger(handle_cmd_error(x, language))
+    return(NULL)
+  }
+
+  if (x$results$resultType == "table") {
+    schema <- data.table::rbindlist(x$results$schema)
+    tbl <- data.table::rbindlist(x$results$data)
+    names(tbl) <- schema$name
+
+    output_tbl <- huxtable::hux(tbl) %>%
+      huxtable::set_all_borders(TRUE) %>%
+      huxtable::set_font_size(10) %>%
+      huxtable::set_position("left")
+
+    huxtable::print_screen(output_tbl)
+    return(NULL)
+  }
+
+  # when result is an image save and present
+  if (x$results$resultType %in% c("images", "image")) {
+    img <- x$results$fileNames[[1]]
+    # read as raw
+    raw <- base64enc::base64decode(what = substr(img, 23, nchar(img)))
+    img <- magick::image_read(raw)
+    grid::grid.newpage()
+    return(grid::grid.raster(img))
+  }
+
+  # otherwise treat the results as standard output
+  # each language needs its own special treatment
+  out <- x$results$data
+
+  # if that output is HTML render via htmltools
+  if (grepl(pattern = "<html|<div", out)) {
+    htmltools::html_print(htmltools::HTML(out))
+    out <- NULL
+  }
+
+  out
+
+}
+
+handle_cmd_error <- function(x, language) {
+  summary <- x$results$summary
+  cause <- x$results$cause
+
+  if (language %in% c("py", "sh")) {
+    msg <- cause
+  }
+
+  if (language == "r") {
+    if (grepl("DATABRICKS_CURRENT_TEMP_CMD__", cause)) {
+      msg <- substring(cause, 62)
+    } else {
+      msg <- cause
+    }
+  }
+
+  if (language %in% c("sql", "scala")) {
+    msg <- summary
+  }
+
+  trimws(msg)
 }