+++ tests/testthat/helper-skip.R
@@ -41,12 +41,9 @@ skip_unless_credentials_set <- function() {
 }
 
 skip_without_venv <- function(env) {
-
   env_available <- reticulate::virtualenv_exists(env)
 
   if (!env_available) {
     skip("Test only runs when venv available")
   }
 }
-
-