+++ NEWS.md
@@ -1,3 +1,14 @@
+# brickster 0.2.5
+
+* Adding `db_repl()` a remote REPL to a Databricks cluster (#53)
+* Removing defunct RStudio add-in for browsing Databricks compute
+* Changes to DESCRIPTION file in preperation for CRAN (#64)
+* Removal of `notebook_use_posit_repo()` and `notebook_enable_htmlwidgets()`
+as they are incompatible with CRAN (#64)
+* Removing kntir engine due to many render edge cases not being solvable
+* Adding shortcuts for REPL under addins
+* Added `db_context_command_run_and_wait`
+
 # brickster 0.2.4
 
 * `open_workspace()` and the rstudio connection pane have been heavily revised